# Specialize Primitive Cursor -- make sorts / merges on a single primitive column faster